Split pointcloud tests into two files
