+++ tests/point_cloud/fixtures.py
@@ -1,0 +1,30 @@
+import subprocess
+import pytest
+
+
+# using a fixture instead of a skipif decorator means we get one aggregated skip
+# message rather than one per test
+@pytest.fixture(scope="session")
+def requires_pdal():
+    has_pdal = False
+    try:
+        import pdal
+
+        assert pdal.Pipeline
+        has_pdal = True
+    except ModuleNotFoundError:
+        pass
+
+    pytest.helpers.feature_assert_or_skip(
+        "pdal package installed", "KART_EXPECT_PDAL", has_pdal, ci_require=False
+    )
+
+
+@pytest.fixture(scope="session")
+def requires_git_lfs():
+    r = subprocess.run(["git", "lfs", "--version"])
+    has_git_lfs = r.returncode == 0
+
+    pytest.helpers.feature_assert_or_skip(
+        "Git LFS installed", "KART_EXPECT_GIT_LFS", has_git_lfs, ci_require=False
+    )
